Refactor to single top-level function per file
