+++ docs-viewer/src/lib/doc-link-types.ts
@@ -1,0 +1,12 @@
+import type { DocFile } from "@/types/docs";
+
+export interface DocLinkTarget {
+  sectionId: string;
+  fileId: string;
+  title: string;
+}
+
+export interface AdjacentDocs {
+  previous?: DocFile;
+  next?: DocFile;
+}
